offline reason and offline cause debug

+++ admin/offline-reason/index.php
@@ -45,7 +45,7 @@ if (!$result) {
                     <td><?php echo date('M d, Y', strtotime($data['date_created'])); ?></td>
                     <td>
                       <a href="edit.php?id=<?php echo $data['reason_id']; ?>" class="btn btn-danger btn-sm b-width">Edit</a>
-                      <a href="delete.php?id=<?php echo $data['reason_id']; ?>" class="btn btn-danger btn-sm b-width">Delete</a>
+                      <a href="delete.php?id=<?php echo $data['reason_id']; ?>" class="btn btn-danger btn-sm b-width" onclick="return confirm('Are you sure you want to delete?');">Delete</a>
                     </td>
                   </tr>
                 <?php
